Parser was first splited
